requirements.txt
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
Brotli==1.0.9
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
DateTime==5.1
decorator==5.1.1
dj-database-url==2.0.0
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
executing==1.2.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
ipython==8.14.0
jedi==0.18.2
jsonschema==4.17.3
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.15.1
PyJWT==2.7.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.7.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
wcwidth==0.2.6
whitenoise==6.5.0
zope.interface==6.0
